requirements.txt
Mezzanine==4.2.3
mezzanine-api
django-scheduler